requirements.txt
Django==4.2.9
djangorestframework==3.14.0
django-cors-headers==4.3.1
psycopg2-binary==2.9.9
python-decouple==3.8
Pillow==10.1.0
django-filter==23.5
razorpay==1.4.1
redis==5.0.1
django-redis==5.4.0
celery==5.3.4
django-celery-beat==2.5.0
requests==2.31.0
stripe==7.9.0
twilio==8.11.0
sendgrid==6.11.0
gunicorn==21.2.0
whitenoise==6.6.0
dj-database-url==2.1.0
django-crispy-forms==2.3
crispy-bootstrap5==2025.6
django-environ==0.11.2
setuptools>=65.0.0
